User y Reservations listos con su respectiva ER, adicionalmente estan relacionados con el Frontend para las consultas segun el caso

+++ src/main/resources/static/js/functionScriptReserva.js
@@ -1,55 +1,91 @@
 function agregarReserva() {
-    //capturar los valores contenidos en las cajas de los input del formualrio index.html
     var datos={
-        id:$("#ReservaID").val(),
         room:$("#room").val(),
-        startdate:$("#startdate").val(),
-        enddate:$("#enddate").val(),
+        dateGetInto:$("#startdate").val(),
+        dateGetOut:$("#enddate").val(),
+        user:{id: +$("#select-user").val()}
     }
-    //convertir a JSON
     let datosPeticion=JSON.stringify(datos);
-
-    //Realizamos la peticion AJAX
     $.ajax({
-        url:"https://g7f0671950df122-jsbm.adb.us-phoenix-1.oraclecloudapps.com/ords/admin/reserva/reserva",
+        url:`http://localhost:8080/api/reservation/new`,
         data:datosPeticion,
         type:"POST",
         contentType:"application/JSON",
-        //si es correcto muestreme mensaje        
-        success:function (respuesta){
+        success:function (){
             console.log("insertado exitosamente")
+            Swal.fire({
+                position: 'center',
+                icon: 'success',
+                title: 'Reservado exitosamente',
+                showConfirmButton: false,
+                timer: 1800
+            });
+            $("#startdate").val("");
+            $("#enddate").val("");
         },
-        //si es incorrecto muestreme error = xhr
         error:function(xhr, status){
             console.log(status);
         }
     });
 }
 
-function listarReserva() {
-    $("#btn-agregarReserva").hide();
-
-    $.ajax({
-        url:"https://g7f0671950df122-jsbm.adb.us-phoenix-1.oraclecloudapps.com/ords/admin/reserva/reserva",
-        Type:"GET",
-        dataType:"json",
-
-        success:function (respuesta) {
-            console.log(respuesta);
-            listarRespuestaReserva(respuesta.items);
-        },
-
-        error:function (xhr, status) {
-            console.log(status);
-        }
-        
-    });
+function validateReservation() {
+    var room = $('#room').val()
+    var dateGetInto = $('#startdate').val()
+    var dateGetOut = $('#enddate').val()
+    if (room.length==0 || dateGetInto.length==0 || dateGetOut.length==0) {
+            Swal.fire({
+                position: 'center',
+                icon: 'warning',
+                title: 'Todos los campos son necesarios!',
+                showConfirmButton: false,
+                timer: 1900
+            })
+    } else {
+        document.formSaveReservation.submit();
+    }
 }
+
+function MisReservas() {
+    $.ajax({
+        url:`http://localhost:8080/api/user/username`,
+        Type:'GET',
+        dataType:'JSON',
+        success:function(request) {
+            console.log(request)
+            listarRespuestaReserva(request.reservations);
+            // listarRespuestaReserva(respuesta);
+        },
+        error:function() {
+            Swal.fire({
+                position: 'center',
+                icon: 'warning',
+                title: 'Todavia no tienes ninguna Reserva activa',
+                showConfirmButton: true,
+            })
+        }
+    }) 
+}
+
+// function listarReservas(id) {
+//     // $("#btn-agregarReserva").hide();
+//     $.ajax({
+//         url:`http://localhost:8080/api/user/reservations/${id}`,
+//         Type:"GET",
+//         dataType:"json",
+//         success:function (respuesta) {
+//             console.log(respuesta);
+//             listarRespuestaReserva(respuesta);
+//         },
+//         error:function (xhr, status) {
+//             console.log(status);
+//         }
+//     });
+// }
 
 function listarRespuestaReserva(items) {
     var tabla=`<table border="2">
                 <tr>
-                    <th>CODIGO</th>
                     <th>HABITACION</th>
                     <th>LLEGADA</th>
                     <th>SALIDA</th>
@@ -57,20 +93,15 @@ function listarRespuestaReserva(items) {
                 // <th colspan="2">ACCIONES</th>
     for (let i = 0; i < items.length; i++) {
         tabla+=`<tr>
-                    <td>${items[i].id} </td>
                     <td>${items[i].room} </td>
-                    <td>${items[i].startdate} </td>
-                    <td>${items[i].enddate} </td>
-                    <td><button onclick="borrarRegistroReserva(${items[i].id})">Cancelar</button></td>
+                    <td>${items[i].dateGetInto} </td>
+                    <td>${items[i].dateGetOut} </td>
+                    <td><button class="btn-borrarRegistroReserva" onclick="borrarRegistroReserva(${items[i].id})">Cancelar</button></td>
                 </tr>`;
     }
     tabla+=`</table>`
-
     console.log(tabla);
-
     $("#listadoReserva").html(tabla)
-    // <td><button onclick="editarRegistroReserva(${items[i].id})">Editar</button></td>
-    // <td><button onclick="borrarRegistroReserva(${items[i].id})">Borrar</button></td>
 }
 
 function editarRegistroReserva(ReservaID) {
@@ -78,30 +109,23 @@ function editarRegistroReserva(ReservaID) {
     $("#btn-agregarReserva").hide();
     $("#btn-listarReserva").hide();
     $("#ReservaID").prop('disabled', true);
-
     var datos={
         id:ReservaID
     }
-
     let datosPeticion=JSON.stringify(datos);
-
     $.ajax({
-        url:"https://g7f0671950df122-jsbm.adb.us-phoenix-1.oraclecloudapps.com/ords/admin/reserva/reserva/"+ReservaID,
+        url:`http://localhost:8080/api/reservation/${ReservaID}`,
         data:datosPeticion,
         type:'GET',
         dataType:'json',
-    
         success:function (respuesta) {
             var items=respuesta.items;
-
-            $('#ReservaID').val(items[0].id),
-            $('#room').val(items[0].room),
-            $('#startdate').val(items[0].startdate),
-            $('#enddate').val(items[0].enddate),
+            $('#ReservaID').val(items.id),
+            $('#room').val(items.room),
+            $('#startdate').val(items.dateGetInto),
+            $('#enddate').val(items.dateGetOut),
             console.log(items);
         },
-
-        //xhr = codigo del error
         error:function(xhr, status){
             console.log(status);
     
@@ -116,25 +140,16 @@ function actualizarRegistroReserva() {
         startdate:$("#startdate").val(),
         enddate:$("#enddate").val(),
     }
-
-    
-
-    //conversion a JSON
     let datosPeticion=JSON.stringify(datos);
-
-    //Hacemos peticion Ajax
     $.ajax({
-        url:"https://g7f0671950df122-jsbm.adb.us-phoenix-1.oraclecloudapps.com/ords/admin/reserva/reserva",
+        url:`http://localhost:8080/api/reservation/update`,
         data:datosPeticion,
         type:'PUT',
         contentType:'application/JSON',
-
         success:function (respuesta) {
             console.log("actualizado!");
             listarReserva();
         },
-
-        //xhr = codigo del error
         error:function(xhr, status){
             console.log(status);
 
@@ -143,26 +158,37 @@ function actualizarRegistroReserva() {
 }
 
 function borrarRegistroReserva(ReservaID) {
-    var datos={
-        id:ReservaID
-    }
-
-    let datosPeticion=JSON.stringify(datos);
-
-    $.ajax({
-        url:"https://g7f0671950df122-jsbm.adb.us-phoenix-1.oraclecloudapps.com/ords/admin/reserva/reserva",
-        data:datosPeticion,
-        type:"DELETE",
-        contentType:"application/JSON",
-
-        success:function (respuesta) {
-            console.log("Borrado");
-            listarReserva();
-        },
-    
-        //xhr = codigo del error
-        error:function(xhr, status){
-            console.log(status);
+    Swal.fire({
+        title: '¿Estas seguro de cancelar esta reserva?',
+        text: "Otra persona puede reservar esta habitacion!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#008F39',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Cancelar de todos modos',
+        cancelButtonText: 'No Cancelar!'
+      }).then((result) => {
+        if (result.isConfirmed) {
+          Swal.fire({
+            title: 'Reserva Cancelada Correctamente',
+            icon: 'success',
+            showConfirmButton: false
+          })
+          $.ajax({
+                url:`http://localhost:8080/api/reservation/delete/${ReservaID}`,
+                type:"DELETE",
+                contentType:"application/JSON",
+                success:function () {
+                    console.log("Borrado");
+                    setTimeout(() => {
+                        document.location.reload();
+                    }, 1800);
+                },
+                error:function(xhr, status){
+                    console.log(status);
+                }
+            });
         }
-    });
+    })
 }
+
